requirements.txt
Flask
gunicorn
Flask-SQLAlchemy
Flask-CORS
python-dotenv
psycopg2-binary




